# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "运维发布管理系统 API 文档",
        "title": "OpsCenter API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@opscenter.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:18080",
    "basePath": "/api",
    "paths": {
...
        "/preprod/scaledown/execute": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据预览 ID 执行预生产环境缩容操作",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "预生产"
                ],
                "summary": "执行预生产缩容",
                "parameters": [
                    {
                        "description": "预览 ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PreviewExecuteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/preprod/scaledown/preview": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "预览预生产环境缩容操作的命令和影响",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "预生产"
                ],
                "summary": "预生产缩容预览",
                "parameters": [
                    {
                        "description": "缩容参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PreprodScaleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/preprod/scaleup/execute": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据预览 ID 执行预生产环境扩容操作",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "预生产"
                ],
                "summary": "执行预生产扩容",
                "parameters": [
                    {
                        "description": "预览 ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PreviewExecuteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/preprod/scaleup/preview": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "预览预生产环境扩容操作的命令和影响",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "预生产"
                ],
                "summary": "预生产扩容预览",
                "parameters": [
                    {
                        "description": "扩容参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PreprodScaleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/preprod/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定服务器的预生产环境资源状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "预生产"
                ],
                "summary": "获取预生产环境状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "服务器 ID",
                        "name": "server_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.PreprodResource"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.PreprodScaleRequest": {
            "type": "object",
            "required": [
                "server_id"
            ],
            "properties": {
                "resource_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "server_id": {
                    "type": "integer"
                }
            }
        },
        "handler.PreviewExecuteRequest": {
            "type": "object",
            "required": [
                "preview_id"
            ],
            "properties": {
                "preview_id": {
                    "type": "string"
                }
            }
        },
...
        "service.PreprodResource": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "string"
                },
                "available": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "current": {
                    "type": "integer"
                },
                "desired": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "target_replicas": {
                    "type": "integer"
                },
                "up_to_date": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}